Add support for explicit API path (#32)
* Upgrade prettier
* Add support for explicit API path

The search endpoint returns an absolute path. The previous version would
prefix `/api/beta` which would break the next URL.

This moves URL generation (that is, combining the baseURL, path, and query vs body params) over to the responsibility of the RequestFactory. This allows for the request factory to append the token to an existing set of query params, as well as formulate new requests that come from a string that is a combination of a path and query param, appending them to the baseURL.

Addresses feedback from #31