Add support for searchStories (#31)
This adds a method to search through stories, including paginating through the results.

Once the results have been fetched, subsequent pages can be fetched by calling `await result.fetchNext()`. This will allow for consumers to not be concerned with how to formulate requests given the format of the pagination data.