Update client to support latest Shortcut v3 API (#184)
* Update client to support latest Shortcut v3 API

In addition to sundry updates, this picks up the renaming of Shortcut's
"Search Stories (Old)" resource to "Query Stories", introducing a
similar change in Api.ts for the respective function.

* Manually add deprecated searchStoriesOld for compat.

This commit adds searchStoriesOld to the client, so that we can release
the addition of queryStories as a backwards-compatible change.